Add two options --createhome and --key
--createhome creates the user's home directory and is largely based on the work and discussion in https://github.com/gregneagle/pycreateuserpkg/pull/14

Changes for this option are to the OptionParser and creation of pkg_data in createuserpkg, make_config_file in userpkg.py and the postinstall script

The postinsstall script now checks for this option and if found runs:  /usr/sbin/createhomedir -c -u $USERNAME

--key will add the user's public key to ~/.ssh/authorized_keys.  Specifying --key implies --createhome.

Changes for this option are in the same locations as --createhome

The postinstall script will add the key to the user's authorized_keys file, creating it (and ~/.ssh) if it doesn't exist.  It will also set permissions on the files.  These commands are run as the user in order to expand ~ and pick up the user's GID when the files are created.  This seems to require the use of eval which means you need to trust the content of the --key variable.  If there is a safer way to do this that would be a good change.